chore: clean up and comment
